Move setup functions to separate file, set up admin user and server name
